Drop Swift 4.2 support
After wrestling with CocoaPods and Travis extensively, I've reached the conclusion that maintaining backwards compatibility with Swift 4.2 while also supporting Swift 5 is difficult, and continuing to maintain automated Swift 4.2 tests to avoid breaking backwards compatibility is also difficult. Since Swift 5 has been out for five months, dropping support for Swift 4.2 in order to reduce the library's maintenance burden is a reasonable tradeoff. If any users of the library are stuck on an old version of Xcode, they can use the prior release of OneTimePassword, or they are welcome to get Swift 4.2 support working and submit a pull request.